+++ springboot-dubbo-server/springboot-dubbo-api/src/main/java/com/example/demo/api/service/UserService.java
@@ -7,5 +7,5 @@ import com.example.demo.api.model.User;
  */
 public interface UserService {
 
-    User findOneById(Integer id);
+    User findOneById(Long id);
 }
