+++ springboot-dubbo-server/springboot-dubbo-provider/src/main/java/com/example/demo/provider/dao/UserMapper.java
@@ -7,5 +7,5 @@ import com.example.demo.api.model.User;
  */
 public interface UserMapper {
 
-    User findOneById(Integer id);
+    User findOneById(Long id);
 }
